Modified Z[sqrt(1)] test sets (HACK)
Modified the Z[sqrt(1)] test set files to have “affix=0” in their header, so they work properly with the APSLQ function. Otherwise the function throws an error. Longer term the APSLQ function needs to be changed to work properly with these affixes, but for now this will suffice to let us run the tests.