Added POSTCHECK
Addd a POSTCHECK callback to the testing code so that additional checking can be performed both before (already possible) and after the main checking. Both PRE and POST check are called in the CHECK procedure.

This required a little bit of re-thinking of the default reporting, which was fixed by a more careful assignment of values to the global CHK variable.