Makefile Revamp
Re-worked the Makefile. The rules are more clearly set up, and it no longer generates impossible output files (output files having gamma_1-gamma but for affixes for which there is no gamma_1).

In addition, the Phase2 testing files have been removed. They will be re-worked and added back in a later commit.