Correct PHASE Selection
Fixed an oversight that the PHASE setting wasn’t properly being set with the new Makefile. This is now fixed.

PHASE is now set explicitly in the generic rules, so Phase specific Makefiles need only set the $(PHASE) Make variable and it will be properly used in the base Makefile.