Phase 2 Testing
Phase 2 testing looks at the required precision. It also introduced the idea of long and short input which allows us to test how well the algorithm performs with more input constants that are needed to find a relation.

Additionally, testing using the LLL algorithm has been added. This necesicated a generalisation of classical integer relations allowing the choice (through setting of a variable) between PSLQ and LLL. Furtheremore, REDUCTION testing has been updated to use this new classical integer relation generalisation.

LLL can produce several possibilities for an integer relation, so the testing infrastructure is modified to look through a set of candidate relations. The best result possible from the set is given.

Finally, added an extra sanity check to each result wherein we attempt to find the integer relation again for one higher precision. This should give a GOOD result if the original result was GOOD (however if the original result was BAD, UNEXPECTED, or FAIL it’s posisble that the sanity-check result might be different; this isn’t considered a problem).