LLL Memory Fix
Changed the LLL_INTEGER_RELATION function to be more memory efficient. A single matrix M is now used that is modified (in place) within the loop (instead of just creating new matrices each iteration). A problem with LLL's remember table necessitated using the forget() function to clear the table each iteration so that the function computed properly on the matrix after its contents were changed.