Snapshot support for tests.
Added a snapshotting capability for stress tests. After each test instance is computed, the results of that computation are written to a snapshot file (currently .BASENAME.sapshot.m where BASENAME is the base name of the output file). The snapshot file is read (if it exists) at the start of computation, and any results already in the file are re-used without recomputation.

In order to achieve this the testing infrastructure needed a not insignificant overhaul. Some redundant code has been removed (moved to stress-test-common.mpl). Moreover, stress-test-common.mpl is now the major driver of the testing; stress-test-PHASE-<N>.mpl implements a function PROCESS_TEST_PROBLEM( ::table ) which is called to test a test instance after it is read from the input file. A callback named CALLCULATE_TEST_PROBLEM( xx, ans, precision ) is provided (by stress-test-common.mpl) which runs TEST, CHECK, etc, and returns a table of output results. Output data is now constructed in-code in a table before beign output to the output file; the table is passed back from CALCULATE_TEST_PROBLEM( ) back through PROCESS_TEST_PROBLEM() to the calling function. Doing this allows a more easy method of adding to the output than was possible when the table was constructed by a frpintf() function at the time of output. The final new function (in stress-test-common.mpl) is PROCESS_LINE( ::string ) which is responsible for the snapshotting. It takes a string which is assumed to be a single line of the input file which it parses, and passes to PROCESS_TEST_PROBLEM(), takes the resulting output table, adds the problem id, and returns it (alternatively returning the literal `EOF` if the string is 0, indicating the end of the file). This function uses 'option: remember;' and is saved to the snapthot file (along with its remember table) after each use, except when `EOF` is returned.