CLASSICAL_INTEGER_RELATION Setup
Moved the check for invalid field into a SETUP() function inside stress-test-CLASSICAL_INTEGER_RELATION.mpl. This way the check happens once, and not on every invocation of TEST( ). In order for this to make sense, we wrapped the main SETUP() call in stress-test-common.mpl inside a try-catch block hwich termintates if an error occurs.