Output System Overhaul
The system for generating output tables was given a complete overhaul. The code introduced with the snapshotting changes (3 commits ago) wasn’t completely implemented, and missed much of the output specific to the APSLQ tests. Furthermore, the chosen method of manipulating table data proved to be prohibitively slow. As a result, the entire method of output generation was re-written.

Now, each function in the execution chain returns a 2-element sequence consisting of a result and a list of output data formatted as one would for creating a table (entries are of the form: index=value). The result is usually one of GOOD, UNEXPECTED, BAD, or FAIL, but in some cases is not (such as the case for TEST() and PRECHECK() which return a set of candidate relations as their result). The function PROCESS_LINE() is the slightly different: it does not return a result at all; it just returns the final list ready for output.

Each function in the chain takes the output list data from the functions it calls, and combines that data with data it generated itself before returning the combined list. The main processing loop in stress-test-common.mpl takes the final data list and prints it to the output file inside a “table( )” string. The table is never actually constructed as a Maple table.

As part of this overhaul, I also took the time to re-name some values, and also added some data to the output of LLL related testing.