Result script fix
Modified the result scripts to work with the new results file format.

Additionally fixed precision-and-timing-results.mpl to properly extract the algebraic integer coefficients.

Finally: modified the temporary file creation in precision-and-timing-results.mpl to work with the version of mktemp that ships with OS X (these modification should also work on Linux).