LLL Update
Updated LLL computation. There are two new abilities.

1. To be able to compute complex integer relations.
2. To output the discriminant(s) of the complex computation corresponding to the relation that is eventually returned.

In order to facilite (2.), the output format of an integer relation computation had to be modified slightly. Each candidate is now of the form [ [relation], [candidate specific output]]. As such APSLQ and PSLQ output needed to be modified appropriately, and the candidate specific output needed to be interleaved with the other output by the CHECK( ) function (in stress-test-common.mpl).