Partial CLASSICAL Testing Results
The results of CLASSICAL testing from the most recent run of tests.

The LLL computations for compex cases with long input (both real and complex constants) are incomplete. It seems that 100 hours was insufficient. Furthermore, these CLASSICAL cases are going to be recomputed anyway after an upcoming commit which will update several of the files which the CLASSICAL output files rely on. These incomplete (albiet predominantly complete) results are committed so that thesis writing can proceed more smoothly while the results are being recomputed after the upcoming update.