Partial Ph2 CLASSICAL Results.
Partial (near complete) results of Phase 2 CLASSICAL testing results from the most recent run of tests. Only the CLASSICAL-LLL results for complex test sets and long output are yet to be completed. These cases have exceeded the maximum computation time on the grid and have consequently ceased computing. They are to be re-submitted to the grid to continue computation (provided that the resumption from the snapshot file does not fail).