Partial Ph2 REDUCTION (LLL)  Results.
Partial (near complete) results of Phase 2 REDUCTION testing (all cases using LLL) from the most recent run of tests. The results for all test sets with large coefficients and long output are yet to be completed. These cases have exceeded the maximum computation time on the grid and have consequently ceased computing. They are to be re-submitted to the grid to continue computation (provided that the resumption from the snapshot file does not fail).