Partial Ph2 REDUCTION (LLL)  Results.
Partial (near complete) results of Phase 2 REDUCTION testing (all cases using LLL) from the most recent run of tests.

All of the complex test sets using real constants appear to be completed, along with a couple of the ones using complex constants. We're still waiting on the remainder of the complex test sets, and the real test sets.