Modified precision and timing results to include LLL Attempts and Computation Result
Modified the precision-and-timing-results.sh to also collate the number of LLL Attempts for LLL computations, and the result of all computations. The former allows us to calculate the value of N for which the lattice finally gave a candidate relation (which was needed in the thesis). The former allows us to filter out BAD and FAIL final results (which will necessarily be reported as 512 decimal digits of precision); this in turn allows our graphed results to be more readable.