Updated count-results.sh to correctly count the new LLL results.
count-results.sh now correctly counts the new LLL results (both CLASSICAL-LLL and REDUCTION-LLL for complex test sets).  Not much needed to change, however we now print empty cells for cases that don’t exist, and also output the key to the cells.

Also fixed up the description of the input parameters, which was incorrect.