Allow Extra Data from PRECHECK Function
Modified PRECHECK and CHECK functions so that the PRECHECK function can give back general data (related to the precheck function itself, or all the candidates checked) in addition to candidate-specific data. This brings it in line with the POSTCHECK function, and more or less all the rest of the functions in the testing infrastructure.