Report results for multiple candidates.
Modified stress-test-common.mpl to check all candidates and store the results in a list, as well as keeping track of the information for the best result. If there was more than one candidate, then the list of results for all candidates is returned as data (labeled by "Candidate Results" in the output file).