Modified stress-test-REDUCTION.mpl.
Modified the REDUCTION stress test file to match the terminology used in the thesis. The tests for these REDUCTION-PSLQ cases are to be rerun on the same hardware as all the other tests (the most recent tests were run on faster hardware, producing potentially disparate results with the rest of the reported results). In light of this planned re-run it seemed sensible to take the opportunity to fix the terminology.