Ph2 REDUCTION (Complex PSLQ) Results.
Complete results of re-run Phase 2 REDUCTION testing (all complex cases using PSLQ) for the most recent modification to the REDUCTION code. This includes reconstruction techniques.

The tests were re-run in order to ensure they were run on the same hardware as all the other tests (due to an oversight the most recent run of complex Phase 2 REDUCTION tests were run on faster hardware than all other tests whose results are stored in this repository and reported in the thesis.