Expunged Phase 1 Results
The Phase 1 results are not reported in the thesis (in which the results in this commit are currently reported) and thus have not been focussed on in any of the recent work. Moreover they have not been updated in some time.

In light of all this they have been expunged so that the results match those reported in the Thesis. The code is all still here, so the Phase 1 tests can be run at a later time with the current testing suite.